refactor: Extract maintenance tabs into separate component
